requirements.txt
web3==6.11.3
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
asyncio==3.4.3
eth-account==0.9.0
eth-utils==2.3.0
cryptography==41.0.7
pycryptodome==3.19.0
colorama==0.4.6
rich==13.7.0
schedule==1.2.0 